+++ cite-gradle-plugin/src/test/fixtures/android-test/src/test/kotlin/com/example/ExampleTest.kt
@@ -1,0 +1,11 @@
+package com.example
+
+import com.jakewharton.cite.__FILE__
+import org.junit.Assert.assertEquals
+import org.junit.Test
+
+class ExampleTest {
+	@Test fun file() {
+		assertEquals("ExampleTest.kt", __FIEELE__)
+	}
+}